django/requirements.txt
Django>=3.2.0,<3.3
python-dotenv>=0.19.0
gunicorn>=20.1.0
PyYAML>=6.0
requests>=2.28.0
